CHROMIUM: rockchip: dmcclk: change parents to pll clocks
This changes the parents of the dmc clk to pll_dpll and pll_gpll from dpll and
gpll. The former parents are just muxes, and we need the actual pll for the dpll
to be the parent. This is because we need to force a recalc rate on the pll_dpll
when the dmc clk changes its rate. We do this by getting the parent.

BUG=chrome-os-partner:31440
TEST=boot jerry

Change-Id: I28d5fa1ff5fd3044ca1cb57f8b5964612508a6f0
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/239311
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 64ccba5ab0827f1db41678b5ab7b179f732a5678)